ci: add -Wall -Wextra -Wpedantic -Wconversion -Wshadow to internal targets (#895)
Introduce strict compiler warning flags for all internal build targets
(ry_lib, ry, ry_tests, native shared libraries) and fix all 239 existing
warnings. Third-party headers (LLVM, GoogleTest) are marked as SYSTEM
includes to suppress their warnings.

Key changes:
- Define RY_WARNING_FLAGS variable and apply per-target via
  target_compile_options(PRIVATE) to avoid leaking into FetchContent
- Replace global include_directories(LLVM) with
  target_include_directories(SYSTEM PUBLIC) on ry_lib
- Add SYSTEM keyword to FetchContent_Declare for GoogleTest (cmake 3.25+)
- Bump cmake_minimum_required from 3.20 to 3.25
- Fix sign-conversion, shorten-64-to-32, unused variables/functions,
  nested anonymous types, and missing field initializers across 51 files

+++ src/codegen_builtin.cpp
@@ -336,14 +336,14 @@ void CodeGen::emitBucketInit(llvm::Value *headerPtr, llvm::StructType *headerTy,
 
     int64_t bucketBytes = initialBucketCount * 8; // sizeof(int64_t)
     llvm::Value *bucketsPtr = builder_.CreateCall(
-        mallocFn, {llvm::ConstantInt::get(i64Ty_, bucketBytes)}, "buckets");
+        mallocFn, {llvm::ConstantInt::get(i64Ty_, static_cast<uint64_t>(bucketBytes))}, "buckets");
     // Fill with 0xFF bytes → each int64_t becomes -1 (EMPTY)
     builder_.CreateCall(memsetFn, {bucketsPtr,
         llvm::ConstantInt::get(i32Ty_, 0xFF),
-        llvm::ConstantInt::get(i64Ty_, bucketBytes)});
+        llvm::ConstantInt::get(i64Ty_, static_cast<uint64_t>(bucketBytes))});
 
     llvm::Value *bcPtr = builder_.CreateStructGEP(headerTy, headerPtr, bucketCountIdx, "bc_ptr");
-    builder_.CreateStore(llvm::ConstantInt::get(i64Ty_, initialBucketCount), bcPtr);
+    builder_.CreateStore(llvm::ConstantInt::get(i64Ty_, static_cast<uint64_t>(initialBucketCount)), bcPtr);
     llvm::Value *bpPtr = builder_.CreateStructGEP(headerTy, headerPtr, bucketsPtrIdx, "bp_ptr");
     builder_.CreateStore(bucketsPtr, bpPtr);
 }
